refactor: new organization for Header and Footer components
