+++ src/components/Footer/Footer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import { Footer } from '../';
+import Footer from './Footer';
 
 describe('<Footer />', () => {
   it('should render properly', () => {
